@wegerk provided some more data which caused problems for the reader.
Those problems have been addressed.

One fairly major change here is that positions and sizes are now
floating point numbers in the data model. This is their data type in
SVG as well. Some rounding logic was added to the writer code in order
to handle clean read/write round trips.

+++ src/pythonGui/karabo_gui/scenemodel/bases.py
@@ -18,13 +18,13 @@ class BaseLayoutModel(BaseSceneObjectData):
     """ A common base class for all layouts
     """
     # The X-coordinate of the layout
-    x = Int
+    x = Float
     # The Y-coordinate of the layout
-    y = Int
+    y = Float
     # The height of the layout
-    height = Int
+    height = Float
     # The width of the layout
-    width = Int
+    width = Float
     # The children of the layout
     children = List(Instance(BaseSceneObjectData))
 
@@ -64,10 +64,10 @@ class BaseWidgetObjectData(BaseSceneObjectData):
     # The possible component type for a parent of the widget (can be empty)
     parent_component = String
     # The X-coordinate of the widget
-    x = Int
+    x = Float
     # The Y-coordinate of the widget
-    y = Int
+    y = Float
     # The height of the widget
-    height = Int
+    height = Float
     # The width of the widget
-    width = Int
+    width = Float
